reused createImageFile function
